# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Data Collection and Payment Management API",
        "description": "Production-ready API for managing users, suppliers, products, collections, rates, and payments with multi-unit tracking, versioned rates, and automated payment calculations. Built on Clean Architecture, SOLID, DRY, and KISS principles.\n *\n * **Authentication:**\n * All protected endpoints require JWT Bearer token authentication. Include the token in the Authorization header.\n *\n * **Rate Limiting:**\n * - Public endpoints (login, register): 5 requests per minute\n * - Standard protected endpoints: 60 requests per minute\n * - Report endpoints: 30 requests per minute (resource intensive)\n *\n * **Version Control:**\n * Update operations on suppliers, products, rates, collections, and payments include version conflict checking to prevent concurrent modification issues. Returns 409 on conflict.\n *\n * **Audit Logging:**\n * All authenticated requests are logged for audit purposes including user actions, IP addresses, and timestamps.",
        "contact": {
            "name": "API Support",
            "email": "support@ledger.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
...
        }
...
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Retrieve a paginated list of products with multi-unit support",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "name",
                                "code",
                                "base_unit",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "description": "Paginated product list",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create new product",
                "description": "Create a new product with multi-unit support",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "base_unit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tea Leaves"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "PROD001"
                                    },
                                    "base_unit": {
                                        "type": "string",
                                        "example": "kg"
                                    },
                                    "supported_units": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "kg",
                                            "g",
                                            "lbs"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "High quality tea leaves"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/products/{id}/current-rate": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get current product rate",
                "description": "Retrieve the current rate for a product based on date and unit",
                "operationId": "getCurrentProductRate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check rate for (defaults to today)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-29"
                        }
                    },
                    {
                        "name": "unit",
                        "in": "query",
                        "description": "Unit to get rate for (defaults to product base_unit)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "kg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "product": {
                                                    "description": "Product details",
                                                    "type": "object"
                                                },
                                                "rate": {
                                                    "description": "Current rate details",
                                                    "type": "object"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "example": "2025-12-29"
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "kg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}/rate-history": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product rate history",
                "description": "Retrieve the complete rate history for a product, optionally filtered by unit",
                "operationId": "getProductRateHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "unit",
                        "in": "query",
                        "description": "Filter history by unit (optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "kg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "product": {
                                                    "description": "Product details",
                                                    "type": "object"
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "kg"
                                                },
                                                "rates": {
                                                    "description": "Rate history",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer Token authentication. Include the token in the Authorization header as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}